Replace GeoAPI namespace with NetTopologySuite.
This was done in a sequence of steps, ignoring the four *Test*.cs files copied in from GeoAPI except where specifically noted:
1. Case-sensitive text find/replace in bulk.  Replacements were:
- "namespace GeoAPI" to "namespace NetTopologySuite"
- "GeoAPI.Geometries." to "NetTopologySuite.Geometries."
- "GeoAPI.IO." to "NetTopologySuite.IO."
- "GeoAPI.GeometryServiceProvider" to "NetTopologySuite.GeometryServiceProvider"
- "GeoAPI.Operation." to "NetTopologySuite.Operation."
- "GeoAPI.Operations." to "NetTopologySuite.Operation." (this fixed some xmldoc that was bad before)

After this, text find for "GeoAPI" did not reveal anything I wanted to do in bulk, so I stopped there.

Doing this caused a lot of formatting / style inconsistencies, so the remainder of this commit was to fix all of that.
Each remaining step seems to have found other related inconsistencies throughout the project, so I elected to keep them.

2. Fix all IDE0001.

3. Fix all IDE0002.

4. Remove all unnecessary "using"s.

5. Sort usings in all files.

6. Manually adjust the four *Test*.cs files copied in from GeoAPI, just enough to get them to build and run as intended (and fix style).
These will be adjusted later to combine them with what NTS started with.